Modularized the LiveReload Configuration
